arranging our app
