+++ src/Components/CTAButton.css
@@ -26,6 +26,7 @@
   .submit-button {
     margin-top: 15px;
     white-space: initial;
+    padding: 0.8rem 2rem;
     /* padding: 0 20px; */
   }
 }
