+++ src/Pages/MediaTracking.css
@@ -4,7 +4,7 @@
 
 .music {
   width: 70%;
-  margin: 1.4rem auto;
+  /* margin: 1.4rem auto; */
   /* background-color: aqua; */
 }
 
@@ -34,11 +34,12 @@
   background-color: #7405a0;
 }
 
-.search-section {
+.m-search-section {
   width: 510px;
   margin: 1.2rem auto;
   /* background-color: aqua; */
 }
+
 .m-search-input {
   border-radius: 8rem;
   background-color: var(--Grey-disabled-02, #f8f7f7);
@@ -50,9 +51,33 @@
   font-weight: 400;
   letter-spacing: -0.5px;
   line-height: 6.4rem;
-  padding: 0.2rem 0.6rem;
+  padding: 0.4rem 0.6rem;
+}
+.m-search-text {
+  font-size: 1.6rem;
 }
 
+.s-b {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.M-category-button {
+  font-size: 1.3rem;
+  font-weight: 600;
+  border: none;
+  border-radius: 6.4rem;
+  padding: 0.8rem 1.8rem;
+  /* background-color: transparent; */
+  background: var(
+    --Button-Primary-button-and-link,
+    linear-gradient(94deg, #b5179e -13.04%, #7209b7 124.22%)
+  );
+  color: var(--background-light-theme-01, #fefefe);
+  cursor: pointer;
+}
 .genre-selector {
   width: 70%;
   margin: 20px auto;
@@ -278,7 +303,7 @@
     top: -4px;
   }
 
-  .search-section {
+  .m-search-section {
     width: 70%;
     margin-top: 5rem;
   }
@@ -370,7 +395,7 @@
     top: -2px;
   }
 
-  .search-section {
+  .m-search-section {
     width: 90%; /* Increased width for better usability */
     margin-top: 3rem;
   }
@@ -396,7 +421,7 @@
     display: none;
   } */
 
-  .search-section {
+  .m-search-section {
     width: 70%;
     margin-left: 4rem;
     /* background-color: #4caf50; */
@@ -405,9 +430,9 @@
 
   .m-search-input {
     /* width: 100%; */
-
-    font-size: 1.5rem;
-    padding: 0.6rem 1rem;
+    gap: 0.1rem;
+    font-size: 1.2rem;
+    padding: 0.4rem 1rem;
   }
 
   /* .search-icon {
@@ -491,14 +516,14 @@
     left: 0;
   }
 
-  .search-section {
+  .m-search-section {
     width: 80%;
     margin-top: 5rem;
   }
 
   .m-search-input {
     font-size: 1.5rem;
-    padding: 0rem 0.4rem;
+    padding: 0.2rem 0.4rem;
   }
 
   /* .genre-selector {
@@ -601,19 +626,19 @@
     font-size: 16px;
   }
 
-  .search-section {
+  .m-search-section {
     width: 360px;
     margin: 1.2rem auto;
   }
 
-  .search-text {
+  .m-search-text {
     font-size: 1.5rem;
     padding: 0.4rem;
   }
 
   .search-icon {
-    width: 2.8rem;
-    height: 2.8rem;
+    width: 2.2rem;
+    height: 2.2rem;
   }
 
   .genre-selector {
@@ -642,13 +667,15 @@
 /* Large devices (desktops, 1024px and up) */
 @media (min-width: 1024px) {
   .music-page {
+    margin-top: 25px;
     flex-direction: row;
   }
 
   .music {
     width: 70%;
-    margin: 1.4rem auto;
+    margin: 1rem auto;
     margin-left: 30rem;
+    /* background-color: #4caf50; */
   }
 
   .back-to-search-container {
@@ -665,14 +692,15 @@
     font-size: 16px;
   }
 
-  .search-section {
+  .m-search-section {
     width: 400px;
-    margin: 1.2rem auto;
+    margin: 0 auto;
+    margin-bottom: 0.6rem;
   }
 
-  .search-text {
+  .m-search-text {
     font-size: 1.5rem;
-    padding: 1.2rem;
+    padding: 0.8rem 1.2rem;
   }
   .genre-selector {
     width: 70%;
