+++ src/Pages/Profile.css
@@ -41,7 +41,9 @@
   border: 3px solid #fff;
   border-radius: 50%;
 }
+
 .profile-name {
+  text-align: left;
   font-size: 2rem;
 }
 
@@ -141,7 +143,7 @@
   gap: 10px;
 }
 
-.btn button {
+.p-button {
   width: 100%;
   margin: 0 auto;
   font-size: 16px;
@@ -152,6 +154,62 @@
   background-color: #fa0000;
   color: #fff;
   cursor: pointer;
+}
+
+.p-modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  backdrop-filter: blur(5px);
+}
+
+.p-modal {
+  background: #fff;
+  padding: 20px 8rem;
+  border-radius: 1.6rem;
+}
+.h2-p {
+  text-align: left;
+  color: #000;
+  font-weight: 600;
+}
+
+/* .p-modal button {
+  margin: 5px;
+} */
+
+.p-modal-btn {
+  /* margin: 20px; */
+  display: flex;
+  flex-direction: row;
+  gap: 3rem;
+  margin-top: 3.2rem;
+}
+
+.primary {
+  padding: 1rem 1.8rem;
+  border-radius: 2rem;
+  font-size: 1.6rem;
+  font-weight: 600;
+  border: none;
+  background: linear-gradient(94deg, #b5179e -13.04%, #7209b7 124.22%);
+  color: #fff;
+}
+
+.sec {
+  padding: 1rem 1.8rem;
+  border-radius: 2rem;
+  font-size: 1.6rem;
+  font-weight: 600;
+  border: 1px solid #881186;
+  color: #000000ce;
+  background-color: transparent;
 }
 
 /* Media Queries */
@@ -204,7 +262,7 @@
     padding: 15px;
   }
 
-  .btn button {
+  .p-button {
     font-size: 1.2rem;
     padding: 0.8rem 4rem;
   }
@@ -262,7 +320,7 @@
     padding: 10px;
   }
 
-  .btn button {
+  .p-button {
     font-size: 10px;
     padding: 0.6rem 3rem;
   }
@@ -371,7 +429,7 @@
     padding: 12px;
   }
 
-  .btn button {
+  .p-button {
     font-size: 10px;
     padding: 0.9rem 6rem;
   }
@@ -422,7 +480,7 @@
     padding: 1.8rem;
   }
 
-  .btn button {
+  .p-button {
     font-size: 1.4rem;
     padding: 1rem 8rem;
   }
@@ -474,7 +532,7 @@
     padding: 18px;
   }
 
-  .btn button {
+  .p-button {
     font-size: 14px;
     padding: 1rem 8rem;
   }
