+++ src/Pages/Profile.js
@@ -1,7 +1,7 @@
-import NavBar from "../Components/NavBar";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { FaApple, FaSpotify, FaSoundcloud } from "react-icons/fa";
 import { useNavigate } from "react-router-dom";
+import NavBar from "../Components/NavBar";
 import "../Pages/Profile.css";
 
 const icons = [
@@ -38,28 +38,37 @@ function Main() {
     </div>
   );
 }
+
 function ProfileName() {
+  const [name, setName] = useState("");
+  const [email, setEmail] = useState("");
+
+  useEffect(() => {
+    setName(localStorage.getItem("userName"));
+    setEmail(localStorage.getItem("userEmail"));
+  }, []);
+
   return (
     <div className="profile-container">
       <div className="profile-details">
         <img
           src="https://images.unsplash.com/photo-1715041348173-1a50eed0ef68?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyN3x8fGVufDB8fHx8fA%3D%3D"
-          alt="name"
+          alt={name}
           className="profile-img"
         />
         <div className="name-flex">
-          <h2 className="profile-name">faychi Favour</h2>
-          <p className="profile-email">faychi@gmail.com</p>
+          <h2 className="profile-name">{name}</h2>
+          <p className="profile-email">{email}</p>
         </div>
       </div>
     </div>
   );
 }
+
 function SyncM() {
   return (
     <div className="syncm">
       <h2>Sync History</h2>
-
       <div className="sync">
         <div className="sync-lay">
           <ul className="plat-icon">
@@ -74,13 +83,13 @@ function SyncM() {
     </div>
   );
 }
+
 function Account() {
   const [form, setForm] = useState("");
   const [password, setPassword] = useState("");
 
   function handleForm(e) {
-    e.preventdefault();
-    // Fix: preventDefault not prevent.default
+    e.preventDefault();
   }
 
   return (
@@ -104,26 +113,47 @@ function Account() {
     </div>
   );
 }
+
 function Button() {
+  const [isModalOpen, setIsModalOpen] = useState(false);
   const history = useNavigate();
 
   function handleDeleteAccount() {
-    // Add logic for deleting account
+    //     // Add logic for deleting account
   }
 
   function handleSignOut() {
-    // Add logic for signing out
-    // Clear user session/token or any other user-related data
     localStorage.removeItem("userToken"); // Assuming user token is stored in localStorage
-
-    // Redirect the user to the login page
+    localStorage.removeItem("userName");
+    localStorage.removeItem("userEmail");
     history("/login");
   }
 
   return (
     <div className="btn">
-      <button onClick={handleDeleteAccount}>Delete Account</button>
-      <button onClick={handleSignOut}>Sign Out</button>
+      <button onClick={handleDeleteAccount} className="p-button">
+        Delete Account
+      </button>
+      <button onClick={() => setIsModalOpen(true)} className="p-button">
+        Sign Out
+      </button>
+
+      {isModalOpen && (
+        <div className="p-modal-overlay">
+          <div className="p-modal">
+            <h2 className="h2-p">Log Out?</h2>
+            <p className="h2-p">Are you sure you want to sign out?</p>
+            <div className="p-modal-btn">
+              <button onClick={handleSignOut} className="primary">
+                Log Out
+              </button>
+              <button onClick={() => setIsModalOpen(false)} className="sec">
+                Dismiss
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
     </div>
   );
 }
