+++ src/Pages/Settings.module.css
@@ -1,0 +1,310 @@
+:root {
+  --background-dark-theme-02: #070615;
+  --text-helper-white-theme-03: #666;
+  --Grey-disabled-02: #f8f7f7;
+  --text-Primary-text-02: #fefefe;
+  --overlay-dark-02: rgba(114, 9, 183, 0.2);
+}
+
+.smain {
+  max-width: 70%;
+  margin: 0 auto;
+}
+
+.header {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 40px;
+  background-color: var(--background-dark-theme-02);
+}
+
+.searchBar {
+  width: 40%;
+  display: flex;
+  align-items: center;
+  background-color: var(--Grey-disabled-02);
+  padding: 6px;
+  border-radius: 80px;
+  font-size: 16px;
+  color: var(--text-helper-white-theme-03);
+}
+
+.searchIcon {
+  width: 18px;
+  margin-right: 8px;
+}
+
+.searchInput {
+  border: none;
+  background: transparent;
+  font-size: 16px;
+  color: var(--text-helper-white-theme-03);
+  outline: none;
+  width: 100%;
+}
+
+.searchInput::placeholder {
+  color: var(--text-helper-white-theme-03);
+}
+
+.navigation {
+  display: flex;
+  justify-content: center;
+  margin-top: 20px;
+  gap: 15px;
+  font-size: 14px;
+  color: var(--text-Primary-text-02);
+  font-weight: 600;
+}
+
+.navItem {
+  padding: 8px 12px;
+  border-radius: 64px;
+  border: 2px solid #b5179e;
+  background-color: transparent;
+  color: var(--text-Primary-text-02);
+  font-size: 14px;
+  font-weight: 600;
+  cursor: pointer;
+}
+
+.navItem:focus {
+  outline: none;
+}
+
+.navItem:hover {
+  background-color: #b5179e;
+}
+
+.highlighted {
+  background: var(--Button-Primary-button-and-link);
+  color: #fff;
+  font-weight: 700;
+}
+
+.content {
+  max-width: 70%;
+  gap: 15px;
+  margin: 30px auto;
+  padding: 16px 37px;
+  background-color: var(--overlay-dark-02);
+  border: 1px solid #b5179e;
+  border-radius: 32px;
+}
+
+.userProfile {
+  display: flex;
+  gap: 20px;
+}
+
+.userImageContainer {
+  position: relative;
+  width: 100px;
+  height: 100px;
+}
+
+.userImage {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+  object-fit: cover;
+  border: 4px solid rgba(254, 254, 254, 1);
+}
+
+.imageOverlay {
+  width: 24px;
+  height: 24px;
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  object-fit: cover;
+  border-radius: 50%;
+}
+
+.userInfo {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  color: var(--text-Primary-text-02);
+}
+
+.userName {
+  font-size: 20px;
+  margin-bottom: 0;
+  font-weight: 600;
+}
+
+.userEmail {
+  font-size: 14px;
+}
+
+.settingItem {
+  display: flex;
+  justify-content: space-between;
+  padding: 10px 0;
+  border-bottom: 1px solid rgba(209, 255, 255, 1);
+}
+
+.settingIconTitle {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  font-size: 20px;
+}
+
+.settingIcon {
+  width: 24px;
+}
+
+.settingTitle {
+  font-size: 15px;
+  font-weight: 400;
+}
+
+.settingAdditionalInfo {
+  font-size: 12px;
+  font-weight: 600;
+}
+
+.logout {
+  display: flex;
+  justify-content: center;
+  margin-top: 20px;
+}
+
+.logoutBtn {
+  padding: 8px 12px;
+  border: 2px solid #b5179e;
+  border-radius: 64px;
+  background-color: transparent;
+  color: var(--text-Primary-text-02);
+  font-weight: 600;
+}
+
+/* Media Queries */
+@media (max-width: 360px) {
+  .smain {
+    max-width: 100%;
+  }
+  .searchBar {
+    width: 100%;
+  }
+  .navigation {
+    max-width: 100%;
+
+    /* flex-direction: column; */
+    margin: 0 auto;
+    /* justify-content: center; */
+    gap: 10px;
+  }
+
+  .navItem {
+    padding: 6px 4px;
+    border-radius: 12px;
+    font-size: 8px;
+  }
+
+  .content {
+    max-width: 75%;
+    padding: 10px 20px;
+  }
+  .userProfile {
+    flex-direction: column;
+    align-items: center;
+  }
+}
+
+@media (max-width: 320px) {
+  .header {
+    padding: 20px;
+  }
+  .navigation {
+    max-width: 100%;
+
+    /* flex-direction: column; */
+    margin: 0 auto;
+    /* justify-content: center; */
+    gap: 10px;
+  }
+
+  .navItem {
+    padding: 6px 4px;
+    border-radius: 12px;
+    font-size: 8px;
+  }
+  .content {
+    padding: 8px 15px;
+  }
+}
+
+@media (min-width: 366px) and (max-width: 767px) {
+  .searchBar {
+    width: 80%;
+    padding: 2px 8px;
+  }
+
+  .searchInput {
+    font-size: 12px;
+    width: 100%;
+  }
+
+  .navigation {
+    max-width: 100%;
+
+    /* flex-direction: column; */
+    margin: 0 auto;
+    /* justify-content: center; */
+    gap: 10px;
+  }
+
+  .navItem {
+    padding: 6px 4px;
+    border-radius: 12px;
+    font-size: 8px;
+  }
+  .content {
+    max-width: 90%;
+    padding: 12px 25px;
+  }
+  .userProfile {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .settingIcon {
+    width: 2.2rem;
+    height: 2rem;
+  }
+
+  .settingTitle {
+    font-size: 12px;
+  }
+}
+
+@media (min-width: 768px) and (max-width: 1023px) {
+  .searchBar {
+    width: 50%;
+  }
+  .navigation {
+    gap: 12px;
+  }
+  .content {
+    max-width: 80%;
+    padding: 14px 30px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .smain {
+    max-width: 70%;
+    margin-left: 20rem;
+  }
+  .searchBar {
+    width: 40%;
+  }
+  .content {
+    max-width: 70%;
+    padding: 16px 37px;
+  }
+}
