+++ src/Pages/faq.module.css
@@ -1,10 +1,21 @@
-.body{
-    display: flex;
+.body {
+  margin-left: 15rem;
+  margin-top: 2rem;
 }
-.main{
-    padding: 5rem 5rem 4rem 4rem;
+.main {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  /* background-color: aqua; */
+  max-width: 65%;
+  padding: 5rem 5rem 4rem 4rem;
+  margin: 0 auto;
+  font-size: 12px;
+  padding: 5rem 5rem 4rem 4rem;
+  height: 80vh;
 }
-.header{
-    margin-top: 0;
-    font-size: 4rem;
+.header {
+  margin-top: 0;
+  font-size: 3rem;
 }
